use of manipulators in C++.
Problem Statement: 
Movie Tickets Sale and Donation to Charity - 
A movie in a local theater is in great demand. To help a local charity, the theater 
owner has decided to donate to the charity a portion of the gross amount generated 
from the movie. This example designs and implements a program that prompts the 
user to input the movie name, adult ticket price, child ticket price, number of adult 
tickets sold, number of child tickets sold, and percentage of the gross amount to be 
donated to the charity. The output of the program is as follows.

Program Analysis: 
Input: The input to the program consists of the movie name, adult ticket price, child ticket 
price, number of adult tickets sold, number of child tickets sold, and percentage of the gross 
amount to be donated to the charity. 
Output: The output is as shown above.