diamond problem(using virtual function)
Problem Statement:
Consider the class hierarchy in the figure below. The class EmployeeCustomer derives
information from both customer and employee classes which in turn derive information from
the class person. Define all the four classes and write a program to create, update and display
the information contained in EmployeeCustomer objects.